requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
Brotli==1.1.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
discord==2.3.2
discord.py==2.4.0
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.20.0
frozenlist==1.4.1
idna==3.7
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
multidict==6.0.5
mutagen==1.47.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pipreqs==0.4.13
platformdirs==4.2.2
prompt_toolkit==3.0.47
pure_eval==0.2.3
pycparser==2.22
pycryptodomex==3.20.0
Pygments==2.18.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32==310
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.1
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
yarg==0.1.9
yarl==1.9.4
yt-dlp==2025.6.30
ffmpeg==1.4
